fixed midleware
