+++ middlewares/index.js
@@ -1,5 +1,7 @@
 const { isEmptyBody } = require("./isEmptyBody");
+const { validationBody } = require('./validationBody');
 
 module.exports = {
-    isEmptyBody
-}
+  isEmptyBody,
+  validationBody,
+};
