+++ middlewares/isEmptyBody.js
@@ -2,7 +2,7 @@ const { HttpError } = require("../helpers/index");
 
 const isEmptyBody = (req, res, next) => {
     if (!Object.keys(req.body).length) {
-      return next(HttpError(400, 'All fields are empty'));
+      return next(HttpError(400, 'missing fields'));
     }
   next();
 }
