+++ models/contacts.json
@@ -70,5 +70,11 @@
     "name": "Alex",
     "email": "britva@gmail.com",
     "phone": "12323424"
+  },
+  {
+    "id": "leQ9bS-y5zOuJeqfrcco_",
+    "name": "Danyiil",
+    "email": "dan@gmail.com",
+    "phone": "13234423"
   }
 ]
