+++ routes/api/contacts.js
@@ -1,17 +1,18 @@
 const express = require('express')
 
 const router = express.Router()
-const contactsController = require("../../controllers/contact-controller");
+const contactsController = require("../../controllers/contact");
 const { isEmptyBody } = require('../../middlewares/index');
+const { validationBody } = require('../../middlewares/index');
 
 
 router.get('/', contactsController.getAllContacts);
 
 router.get('/:contactId', contactsController.getContact);
 
-router.post('/', isEmptyBody ,contactsController.createContact);
+router.post('/', isEmptyBody, validationBody, contactsController.createContact);
 
-router.put('/:contactId', isEmptyBody, contactsController.updateContactById);
+router.put('/:contactId', isEmptyBody, validationBody, contactsController.updateContactById);
 
 router.delete('/:contactId', contactsController.deleteContact);
